move hash executor into new file
